Installation of CentOS 6.7 x86_64 guest OS in Virtuozzo 7 and virtual machine management
Virtuozzo is an virtualization and automation solution built on top of OpenVZ. Virtuozzo provides improvements and additional functionality in the areas of density, management tools, recovery, etc. Specific benefits of Virtuozzo compared to OpenVZ can be found in this article - https://openvz.org/Comparison ( https://openvz.org/Comparison ) .

Virtuozzo is built on top of OpenVZ technologies but provides more features, services, and support.
https://openvz.org/Virtuozzo ( https://openvz.org/Virtuozzo )