remoteu2f
remoteu2f ( http://blitiri.com.ar/p/remoteu2f ) is a project that enables the use of a FIDO U2F ( https://www.yubico.com/applications/fido/ ) security key remotely, through a lightly-trusted proxy.

It is useful mainly to use as a second factor authentication for ssh and sudo-over-ssh.

This asciicast shows how it is set up for the first time, and how ssh into the server afterwards looks like.

https://blitiri.com.ar/p/remoteu2f ( https://blitiri.com.ar/p/remoteu2f )